requirements.txt
python-telegram-bot==20.6
fastapi==0.111.0
uvicorn==0.25.0
yt-dlp==2025.9.6
aiofiles==23.1.0
